refactor(ref-api-02): batch controller slim — extract docs/serializers/validators (PRs 2-7/7)
- wallet/entry_resources.py: inline @doc() kwargs → app/schemas/openapi/wallet/docs.py (397 → 218 LOC)
- dashboard/resources.py: inline @doc() kwargs → app/schemas/openapi/dashboard/docs.py (403 → 222 LOC)
- fiscal/resources.py: inline serializers → app/schemas/openapi/fiscal/response.py (449 → 389 LOC)
- shared_entries/resources.py: parsing helpers → shared_entries/validators.py (469 → 324 LOC)
