+++ app/schemas/openapi/fiscal/__init__.py
@@ -1,0 +1,1 @@
+
