+++ app/schemas/openapi/wallet/__init__.py
@@ -1,0 +1,1 @@
+
